+++ setup_extension.py
@@ -7,17 +7,21 @@
 File created: August 21st 2020
 
 Modified By: hsky77
-Last Updated: September 5th 2020 23:17:55 pm
+Last Updated: October 5th 2020 20:00:12 pm
 '''
 
+import os
 import shutil
 import setuptools
 
 from hyssop import Version as hy_ver
 from hyssop_extension import Version, __name__
 
-shutil.rmtree('dist')
-shutil.rmtree('build')
+if os.path.isdir('dist'):
+    shutil.rmtree('dist')
+
+if os.path.isdir('build'):
+    shutil.rmtree('build')
 
 with open("hyssop_extension/README.md", "r") as fh:
     long_description = fh.read()
